services/data-collector/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
APScheduler==3.10.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
lxml==4.9.3
